+++ CHANGELOG.md
@@ -4,6 +4,34 @@ GovReady-Q Release Notes
 v0.9.1.52 (February 16, 2021)
 ------------------------
 
+**Developer changes**
+
+* Created a ElementEditForm Django form in conjunction with some functional changes to avoid name collisions issues with component library.
+
+**Feature changes**
+
+* Remove a component and its statements from a system.
+
+**UI changes**
+
+* Added a button to system selected component page to remove a component from the system.
+* Moved "Add a component" to a system drop down to top of selected component page.
+* Added a search text box for each searching of components in the library and their statements.
+* Added pagnation to the component library and their statements.
+* Added a reset button for explicit resetting of component search.
+
+**Developer changes**
+
+* Move inclusion of `edit-component-modal.html` from `base.html` to `components/element_detail_tabs.html`.
+* Fix sort control order in `component_library_component` on the `components/element_detail_tabs.html` using the `natsort` package to sort SID correctly.
+
+**Data changes**
+
+* Alter Element description field to be blank and none.
+
+v0.9.1.52 (February 16, 2021)
+-----------------------------
+
 Add System Assessment Report tracking to associate assessments and evidence with the system.
 Add initial dynamic status information to the project page.
 Project page displays mini-dashboard of compliance stats.
@@ -28,7 +56,6 @@ Project page displays mini-dashboard of compliance stats.
 * Properly restrict statement history access to users with system, staff, or admin permissions.
 * Avoid name collisions when cloning a component.
 * Replaced function-based views with class-based listview for SelectedComponentsList, ProjectList.
-* Avoid name collisions when cloning a component.
 * Default to not use Django Debug Toolbar. Added new `enable_tool_bar` parameter option for `local/environment.json` to allow users to enable(True) or disable(False) the Django Debug Toolbar.
 * Adding DummyCache to prevent real caching while running automated tests.
 * Refactored use of random package to use secure secrets module.